+++ src/icons/index.tsx
@@ -1,0 +1,3 @@
+import { SearchSVG } from "./SearchSVG";
+
+export { SearchSVG };
