+++ src/templates/BaseTemplate.tsx
@@ -1,5 +1,5 @@
-import Footer from "../components/footer/Footer";
-import Header from "../components/header/Header";
+import { Footer } from "../components/footer/Footer";
+import { Header } from "../components/header/Header";
 
 type BaseTemplateProps = {
   children: React.ReactNode;
